fix speed to be u8, other things i will need to revert because this is actually just a test commit

+++ crates/game/tataku-game/src/managers/spectator_manager.rs
@@ -354,7 +354,7 @@ impl HostMap {
         map_hash: Md5Hash,
         playmode: String,
         mods: &[ModDefinition],
-        speed: u16
+        speed: u8
     ) -> Self {
         Self { 
             map_hash,
